Write basic start menu code without implementing
The basic code/concept for the start menu user select is now in
startmenu.s file.

The changes to snes.s were returning button register as r0 instead of
r9, which was getting restored anyway.

In main.s, the intro screen image file was updated, and the SNES
controller initialization code was added at the start of the program.