Add main menu, allow for game restore - untested
After checking program specifications, there must be an option mid game
to view a game menu. If start is pressed, then menu is excited and the
current game is resumed.

This posed an issue as there was no current way to redraw the game from
the game state, as it was represented by block/no block, i.e. 0 or 1.
This was changed to be 0-7, where 1-7 represent blocks. Now we can loop
through and depending on the number, draw a different coloured square.

To do this, the game state update code in various places had to be
changed, so will need testing.

Also, the specifications say that the blocks can rotate both ways, but
we will deal with one way first.